move functions to filled post utils file
