+++ utils/ind_cqc_filled_posts_by_job_role_utils/utils.py
@@ -1,5 +1,9 @@
 from pyspark.sql import DataFrame, functions as F
 
+from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
+
+from pyspark.sql import DataFrame, functions as F
+
 from utils.column_names.cleaned_data_files.ascwds_worker_cleaned import (
     AscwdsWorkerCleanedColumns as AWKClean,
 )
@@ -10,6 +14,36 @@ from utils.value_labels.ascwds_worker.ascwds_worker_mainjrid import (
 )
 
 list_of_job_roles = list(AscwdsWorkerValueLabelsMainjrid.labels_dict.values())
+
+
+def count_registered_manager_names(df: DataFrame) -> DataFrame:
+    """
+    Adds a column with a count of elements within list of registered manager names.
+
+    This function uses the size method to count elements in list of strings. This method
+    returns the count, including null elements within a list, when list partially populated.
+    It returns 0 when list is empty. It returns -1 when row is null.
+    Therefore, after the counting, this function recodes values of -1 to 0.
+
+    Args:
+        df (DataFrame): A dataframe containing list of registered manager names.
+
+    Returns:
+        DataFrame: A dataframe with count of elements in list of registered manager names.
+    """
+
+    df = df.withColumn(
+        IndCQC.registered_manager_count, F.size(F.col(IndCQC.registered_manager_names))
+    )
+
+    df = df.withColumn(
+        IndCQC.registered_manager_count,
+        F.when(F.col(IndCQC.registered_manager_count) == -1, F.lit(0)).otherwise(
+            F.col(IndCQC.registered_manager_count)
+        ),
+    )
+
+    return df
 
 
 def count_job_role_per_establishment_as_columns(
